transform-image/requirements.txt
diffusers==0.20.2
transformers==4.32.1
torch~=2.0.1
accelerate==0.22.0
requests~=2.31.0
Pillow~=10.0.0